+++ metric/internal/global/instruments.go
@@ -12,12 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package global // import "go.opentelemetry.io/otel/internal/global"
+package global // import "go.opentelemetry.io/otel/metric/internal/global"
 
 import (
 	"context"
 	"sync/atomic"
 
+	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/embedded"
 	"go.opentelemetry.io/otel/metric/instrument"
@@ -44,7 +45,7 @@ var _ instrument.Float64ObservableCounter = (*afCounter)(nil)
 func (i *afCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableCounter(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -73,7 +74,7 @@ var _ instrument.Float64ObservableUpDownCounter = (*afUpDownCounter)(nil)
 func (i *afUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableUpDownCounter(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -102,7 +103,7 @@ var _ instrument.Float64ObservableGauge = (*afGauge)(nil)
 func (i *afGauge) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableGauge(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -131,7 +132,7 @@ var _ instrument.Int64ObservableCounter = (*aiCounter)(nil)
 func (i *aiCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableCounter(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -160,7 +161,7 @@ var _ instrument.Int64ObservableUpDownCounter = (*aiUpDownCounter)(nil)
 func (i *aiUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableUpDownCounter(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -189,7 +190,7 @@ var _ instrument.Int64ObservableGauge = (*aiGauge)(nil)
 func (i *aiGauge) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableGauge(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -217,7 +218,7 @@ var _ instrument.Float64Counter = (*sfCounter)(nil)
 func (i *sfCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64Counter(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -243,7 +244,7 @@ var _ instrument.Float64UpDownCounter = (*sfUpDownCounter)(nil)
 func (i *sfUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64UpDownCounter(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -269,7 +270,7 @@ var _ instrument.Float64Histogram = (*sfHistogram)(nil)
 func (i *sfHistogram) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64Histogram(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -295,7 +296,7 @@ var _ instrument.Int64Counter = (*siCounter)(nil)
 func (i *siCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64Counter(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -321,7 +322,7 @@ var _ instrument.Int64UpDownCounter = (*siUpDownCounter)(nil)
 func (i *siUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64UpDownCounter(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -347,7 +348,7 @@ var _ instrument.Int64Histogram = (*siHistogram)(nil)
 func (i *siHistogram) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64Histogram(i.name, i.opts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		otel.Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)